// apps/keycloak-theme/src/kc.gen.tsx
// This file is auto-generated by the `update-kc-gen` command. Do not edit it manually.
// Hash: e05ee4b974953dfc0a3b5d01cc0f74c77092ee9be1ee91be05e02718cb84150d


/* eslint-disable */

// @ts-nocheck

// noinspection JSUnusedGlobalSymbols

import { lazy, Suspense, type ReactNode } from "react";

export type ThemeName = "rav5";

export const themeNames: ThemeName[] = ["rav5"];

export type KcEnvName = "RAV5_DEMO_USER_EMAIL" | "RAV5_DEMO_USER_PASSWORD" | "RAV5_DEMO_ADMIN_EMAIL" | "RAV5_DEMO_ADMIN_PASSWORD" | "RAV5_PUBLIC_SITE_URL" | "RAV5_APP_URL";

export const kcEnvNames: KcEnvName[] = ["RAV5_DEMO_USER_EMAIL", "RAV5_DEMO_USER_PASSWORD", "RAV5_DEMO_ADMIN_EMAIL", "RAV5_DEMO_ADMIN_PASSWORD", "RAV5_PUBLIC_SITE_URL", "RAV5_APP_URL"];

export const kcEnvDefaults: Record<KcEnvName, string> = {
  "RAV5_DEMO_USER_EMAIL": "",
  "RAV5_DEMO_USER_PASSWORD": "",
  "RAV5_DEMO_ADMIN_EMAIL": "",
  "RAV5_DEMO_ADMIN_PASSWORD": "",
  "RAV5_PUBLIC_SITE_URL": "",
  "RAV5_APP_URL": ""
};

/**
 * NOTE: Do not import this type except maybe in your entrypoint. 
 * If you need to import the KcContext import it either from src/login/KcContext.ts or src/account/KcContext.ts.
 * Depending on the theme type you are working on.
 */
export type KcContext =
    | import("./login/KcContext").KcContext
    ;

declare global {
    interface Window {
        kcContext?: KcContext;
    }
}

export const KcLoginPage = lazy(() => import("./login/KcPage"));

export function KcPage(
    props: {
        kcContext: KcContext;
        fallback?: ReactNode;
    }
) {
    const { kcContext, fallback } = props;
    return (
        <Suspense fallback={fallback}>
            {(() => {
                switch (kcContext.themeType) {
                    case "login": return <KcLoginPage kcContext={kcContext} />;
                }
            })()}
        </Suspense>
    );
}

// NOTE: This is exported here only because in Webpack environnement it works differently
export const BASE_URL = import.meta.env.BASE_URL